+++ packages/main/src/TableHeaderRowTemplate.tsx
@@ -16,7 +16,6 @@ export default function TableHeaderRowTemplate(this: TableHeaderRow, ariaColInde
 					aria-description={this._selectionCellAriaDescription}
 					aria-colindex={ariaColIndex++}
 					data-ui5-table-selection-cell
-					data-ui5-table-cell-fixed
 					data-ui5-acc-text=""
 				>
 					{ !this._isMultiSelect ?